move urllib_retrieve_hook to `_util.py`
